+++ internal/api/cron/func_create.go
@@ -13,21 +13,21 @@ type createRequest struct {
 	Name                string `form:"name" binding:"required"`           // 任务名称
 	Spec                string `form:"spec" binding:"required"`           // crontab 表达式
 	Command             string `form:"command" binding:"required"`        // 执行命令
-	Protocol            int32  `form:"protocol" binding:"required"`       // 执行方式 1:shell 2:http
-	HttpMethod          int32  `form:"http_method"`                       // http 请求方式 1:get 2:post
-	Timeout             int32  `form:"timeout" binding:"required"`        // 超时时间(单位:秒)
-	RetryTimes          int32  `form:"retry_times" binding:"required"`    // 重试次数
-	RetryInterval       int32  `form:"retry_interval" binding:"required"` // 重试间隔(单位:秒)
-	NotifyStatus        int32  `form:"notify_status" binding:"required"`  // 执行结束是否通知 1:不通知 2:失败通知 3:结束通知 4:结果关键字匹配通知
-	NotifyType          int32  `form:"notify_type"`                       // 通知类型 1:邮件 2:webhook
+	Protocol            int64  `form:"protocol" binding:"required"`       // 执行方式 1:shell 2:http
+	HttpMethod          int64  `form:"http_method"`                       // http 请求方式 1:get 2:post
+	Timeout             int64  `form:"timeout" binding:"required"`        // 超时时间(单位:秒)
+	RetryTimes          int64  `form:"retry_times" binding:"required"`    // 重试次数
+	RetryInterval       int64  `form:"retry_interval" binding:"required"` // 重试间隔(单位:秒)
+	NotifyStatus        int64  `form:"notify_status" binding:"required"`  // 执行结束是否通知 1:不通知 2:失败通知 3:结束通知 4:结果关键字匹配通知
+	NotifyType          int64  `form:"notify_type"`                       // 通知类型 1:邮件 2:webhook
 	NotifyReceiverEmail string `form:"notify_receiver_email"`             // 通知者邮箱地址(多个用,分割)
 	NotifyKeyword       string `form:"notify_keyword"`                    // 通知匹配关键字(多个用,分割)
 	Remark              string `form:"remark"`                            // 备注
-	IsUsed              int32  `form:"is_used" binding:"required"`        // 是否启用 1:是  -1:否
+	IsUsed              int64  `form:"is_used" binding:"required"`        // 是否启用 1:是  -1:否
 }
 
 type createResponse struct {
-	Id int32 `json:"id"` // 主键ID
+	Id int64 `json:"id"` // 主键ID
 }
 
 // Create 创建任务
