+++ internal/api/cron/func_updateused.go
@@ -10,11 +10,11 @@ import (
 
 type updateUsedRequest struct {
 	Id   string `form:"id"`   // 主键ID
-	Used int32  `form:"used"` // 是否启用 1:是 -1:否
+	Used int64  `form:"used"` // 是否启用 1:是 -1:否
 }
 
 type updateUsedResponse struct {
-	Id int32 `json:"id"` // 主键ID
+	Id int64 `json:"id"` // 主键ID
 }
 
 // UpdateUsed 更新任务为启用/禁用
@@ -52,7 +52,7 @@ func (h *handler) UpdateUsed() core.HandlerFunc {
 			return
 		}
 
-		id := int32(ids[0])
+		id := int64(ids[0])
 
 		err = h.cronService.UpdateUsed(ctx, id, req.Used)
 		if err != nil {
