+++ internal/repository/cron/cron_remove_task.go
@@ -2,7 +2,7 @@ package cron
 
 import "github.com/spf13/cast"
 
-func (s *server) RemoveTask(taskId int) {
+func (s *server) RemoveTask(taskId int64) {
 	name := cast.ToString(taskId)
 	s.cron.RemoveJob(name)
 }
